+++ src/components/PinField/PinField.tsx
@@ -1,0 +1,12 @@
+import PinFieldControlled from "./PinFieldControlled";
+import PinFieldUncontrolled from "./PinFieldUncontrolled";
+import type * as T from "./PinField.types";
+
+const PinField = (props: T.Props) => {
+	const { value } = props;
+
+	if (value !== undefined) return <PinFieldControlled {...props} />;
+	return <PinFieldUncontrolled {...props} />;
+};
+
+export default PinField;