+++ CHANGELOG.md
@@ -7,11 +7,14 @@ Features:
 - Toast: Added support for toast render boundaries
 - useHotkeys: disable flag
 - Modal: overlayClassName
+- PinField: new component
+- useHotkeys: preventDefault flag
 
 Fixes:
 
 - View: Fixed dividers incorrectly working with React.Fragments
 - useHotkeys: auto preventDefault
+- useHotkeys: fixed used hotkeys cleanup
 - DropdownMenu: fixed double click handling
 - Slider: Float values
 - Slider: Tabular numbers